requirements.txt
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1727892028500/work
asttokens==2.4.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1722977137225/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1715091184126/work
Brotli @ file:///D:/bld/brotli-split_1725267609074/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1724028158384/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
cffi @ file:///D:/bld/cffi_1725560655177/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1728479282467/work
click @ file:///D:/bld/click_1692312014553/work
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing==2.0.1
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1697791558612/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1711991025291/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1726459485162/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1729190692267/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1717461632069/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1719171233697/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1725921340658/work
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1724334859652/work
jedi==0.19.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1720529478715/work
jsonschema-specifications @ file:///tmp/tmpvslgxhz5/src
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1724331334887/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1686175045316/work
MarkupSafe @ file:///D:/bld/markupsafe_1729351305211/work
matplotlib-inline==0.1.7
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1704317613764/work
moviepy @ file:///home/conda/feedstock_root/build_artifacts/moviepy_1665160419595/work
narwhals @ file:///home/conda/feedstock_root/build_artifacts/narwhals_1729515497107/work
nest-asyncio==1.6.0
numpy @ file:///D:/bld/numpy_1707225519803/work/dist/numpy-1.26.4-cp312-cp312-win_amd64.whl#sha256=ffaa9485afb654c70a658c3ad4b7d2e60c60d8450d3779408d3a496564dc1d43
packaging==24.1
pandas @ file:///D:/bld/pandas_1726878552024/work
parso==0.8.4
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow @ file:///D:/bld/pillow_1726075241382/work
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1694617248815/work
platformdirs==4.2.2
pluggy==1.5.0
proglog @ file:///home/conda/feedstock_root/build_artifacts/proglog_1728452852274/work
prompt_toolkit==3.0.47
protobuf @ file:///D:/bld/protobuf_1727108932349/work/bazel-bin/python/dist/protobuf-5.27.5-cp312-abi3-win_amd64.whl#sha256=8906bbb61809c1399bf1070d66276462f83bf4a63723e4b743ebd3f01ab573b2
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyClamd==0.4.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1667589451974/work
Pygments==2.18.0
PySocks @ file:///D:/bld/pysocks_1661604991356/work
pytest==8.3.3
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-magic-bin==0.4.14
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
pywin32==306
PyYAML @ file:///D:/bld/pyyaml_1725456271438/work
pyzmq==26.0.3
referencing @ file:///home/conda/feedstock_root/build_artifacts/referencing_1714619483868/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1717057054362/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich_1729622917073/work/dist
rpds-py @ file:///D:/bld/rpds-py_1725327174158/work
setuptools==75.1.0
six==1.16.0
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1634310307496/work
stack-data==0.6.3
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1727818352917/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1722278858289/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tornado==6.4.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1722737464726/work
traitlets==5.14.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1727140567071/work
tzlocal @ file:///D:/bld/tzlocal_1724952188353/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1726496430923/work
validators @ file:///home/conda/feedstock_root/build_artifacts/validators_1725433541799/work
watchdog @ file:///D:/bld/watchdog_1727641431273/work
wcwidth==0.2.13
wheel==0.44.0
widgetsnbextension @ file:///home/conda/feedstock_root/build_artifacts/widgetsnbextension_1724331337528/work
win_inet_pton @ file:///D:/bld/win_inet_pton_1727796272493/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1726248574750/work
zstandard==0.23.0
